requirements.txt
fastapi==0.128.6
uvicorn==0.40.0
httpx==0.28.1
aiofiles==25.1.0
python-dotenv==1.1.0
mcp==1.26.0
mcp-agent==0.2.6
redis==5.2.1
google-genai==1.3.0
openai==1.63.2
